User Authentication and goal protection(using JWT) done!

+++ backend/controller/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
     console.log("GET successfully sent !")
@@ -23,7 +24,8 @@ const postGoal = asyncHandler(async (req, res) => {
     }
     else{
         const goal = await Goal.create({
-            text: req.body.text
+            text: req.body.text,
+            user: req.user.id,
         })
         res.status(200).json(goal)
     }
@@ -41,8 +43,22 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found !')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized !')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
+            new: true,
     })
 
     res.status(200).json(updatedGoal)
@@ -60,6 +76,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found !')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized !')
     }
 
     await goal.deleteOne()
